LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookies/name/name-ctl-expected.txt


PASS Cookie with %x0 in name is truncated.
FAIL Cookie with %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test10" but got "test10; name=10"
PASS Cookie with %xd in name is truncated.
FAIL Cookie with %x1 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test1"
FAIL Cookie with %x2 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test2"
FAIL Cookie with %x3 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test3"
FAIL Cookie with %x4 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test4"
FAIL Cookie with %x5 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test5"
FAIL Cookie with %x6 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test6"
FAIL Cookie with %x7 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test7"
FAIL Cookie with %x8 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test8"
FAIL Cookie with %x9 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test9"
FAIL Cookie with %xb in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test11"
FAIL Cookie with %xc in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test12"
FAIL Cookie with %xe in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test14"
FAIL Cookie with %xf in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test15"
FAIL Cookie with %x10 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test16"
FAIL Cookie with %x11 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test17"
FAIL Cookie with %x12 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test18"
FAIL Cookie with %x13 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test19"
FAIL Cookie with %x14 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test20"
FAIL Cookie with %x15 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test21"
FAIL Cookie with %x16 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test22"
FAIL Cookie with %x17 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test23"
FAIL Cookie with %x18 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test24"
FAIL Cookie with %x19 in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test25"
FAIL Cookie with %x1a in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test26"
FAIL Cookie with %x1b in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test27"
FAIL Cookie with %x1c in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test28"
FAIL Cookie with %x1d in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test29"
FAIL Cookie with %x1e in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test30"
FAIL Cookie with %x1f in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test31"
FAIL Cookie with %x7f in name is rejected. assert_equals: The cookie was rejected. expected "" but got "test127name=127"
PASS Cookie with %x1 after %x0 in name is truncated.
PASS Cookie with %x2 after %x0 in name is truncated.
PASS Cookie with %x3 after %x0 in name is truncated.
PASS Cookie with %x4 after %x0 in name is truncated.
PASS Cookie with %x5 after %x0 in name is truncated.
PASS Cookie with %x6 after %x0 in name is truncated.
PASS Cookie with %x7 after %x0 in name is truncated.
PASS Cookie with %x8 after %x0 in name is truncated.
PASS Cookie with %x9 after %x0 in name is truncated.
PASS Cookie with %xb after %x0 in name is truncated.
PASS Cookie with %xc after %x0 in name is truncated.
PASS Cookie with %xe after %x0 in name is truncated.
PASS Cookie with %xf after %x0 in name is truncated.
PASS Cookie with %x10 after %x0 in name is truncated.
PASS Cookie with %x11 after %x0 in name is truncated.
PASS Cookie with %x12 after %x0 in name is truncated.
PASS Cookie with %x13 after %x0 in name is truncated.
PASS Cookie with %x14 after %x0 in name is truncated.
PASS Cookie with %x15 after %x0 in name is truncated.
PASS Cookie with %x16 after %x0 in name is truncated.
PASS Cookie with %x17 after %x0 in name is truncated.
PASS Cookie with %x18 after %x0 in name is truncated.
PASS Cookie with %x19 after %x0 in name is truncated.
PASS Cookie with %x1a after %x0 in name is truncated.
PASS Cookie with %x1b after %x0 in name is truncated.
PASS Cookie with %x1c after %x0 in name is truncated.
PASS Cookie with %x1d after %x0 in name is truncated.
PASS Cookie with %x1e after %x0 in name is truncated.
PASS Cookie with %x1f after %x0 in name is truncated.
PASS Cookie with %x7f after %x0 in name is truncated.
FAIL Cookie with %x1 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test1term" but got "na"
FAIL Cookie with %x2 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test2term" but got "na"
FAIL Cookie with %x3 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test3term" but got "na"
FAIL Cookie with %x4 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test4term" but got "na"
FAIL Cookie with %x5 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test5term" but got "na"
FAIL Cookie with %x6 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test6term" but got "na"
FAIL Cookie with %x7 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test7term" but got "na"
FAIL Cookie with %x8 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test8term" but got "na"
FAIL Cookie with %x9 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test9term" but got "na"
FAIL Cookie with %xb after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test11term" but got "na"
FAIL Cookie with %xc after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test12term" but got "na"
FAIL Cookie with %xe after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test14term" but got "na"
FAIL Cookie with %xf after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test15term" but got "na"
FAIL Cookie with %x10 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test16term" but got "na"
FAIL Cookie with %x11 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test17term" but got "na"
FAIL Cookie with %x12 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test18term" but got "na"
FAIL Cookie with %x13 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test19term" but got "na"
FAIL Cookie with %x14 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test20term" but got "na"
FAIL Cookie with %x15 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test21term" but got "na"
FAIL Cookie with %x16 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test22term" but got "na"
FAIL Cookie with %x17 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test23term" but got "na"
FAIL Cookie with %x18 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test24term" but got "na"
FAIL Cookie with %x19 after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test25term" but got "na"
FAIL Cookie with %x1a after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test26term" but got "na"
FAIL Cookie with %x1b after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test27term" but got "na"
FAIL Cookie with %x1c after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test28term" but got "na"
FAIL Cookie with %x1d after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test29term" but got "na"
FAIL Cookie with %x1e after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test30term" but got "na"
FAIL Cookie with %x1f after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test31term" but got "na"
FAIL Cookie with %x7f after %xa in name is truncated. assert_equals: The cookie was set as expected. expected "test127term" but got "test127term; name=127"
PASS Cookie with %x1 after %xd in name is truncated.
PASS Cookie with %x2 after %xd in name is truncated.
PASS Cookie with %x3 after %xd in name is truncated.
PASS Cookie with %x4 after %xd in name is truncated.
PASS Cookie with %x5 after %xd in name is truncated.
PASS Cookie with %x6 after %xd in name is truncated.
PASS Cookie with %x7 after %xd in name is truncated.
PASS Cookie with %x8 after %xd in name is truncated.
PASS Cookie with %x9 after %xd in name is truncated.
PASS Cookie with %xb after %xd in name is truncated.
PASS Cookie with %xc after %xd in name is truncated.
PASS Cookie with %xe after %xd in name is truncated.
PASS Cookie with %xf after %xd in name is truncated.
PASS Cookie with %x10 after %xd in name is truncated.
PASS Cookie with %x11 after %xd in name is truncated.
PASS Cookie with %x12 after %xd in name is truncated.
PASS Cookie with %x13 after %xd in name is truncated.
PASS Cookie with %x14 after %xd in name is truncated.
PASS Cookie with %x15 after %xd in name is truncated.
PASS Cookie with %x16 after %xd in name is truncated.
PASS Cookie with %x17 after %xd in name is truncated.
PASS Cookie with %x18 after %xd in name is truncated.
PASS Cookie with %x19 after %xd in name is truncated.
PASS Cookie with %x1a after %xd in name is truncated.
PASS Cookie with %x1b after %xd in name is truncated.
PASS Cookie with %x1c after %xd in name is truncated.
PASS Cookie with %x1d after %xd in name is truncated.
PASS Cookie with %x1e after %xd in name is truncated.
PASS Cookie with %x1f after %xd in name is truncated.
PASS Cookie with %x7f after %xd in name is truncated.
